Remove totally unneeded and dangerous deleteBucket() from AmazonS3FileBackend.
1) The very point of doCleanInternal() is to remove directories when all files
inside of them are deleted. E.g. thumb directory when all thumbnails are gone.

In Amazon S3, directories are virtual, so there is nothing to do.
This function must be a no-op.

2) The call to deleteBucket() is a bug.

There is absolutely no point to call deleteBucket(), given that buckets are
constant (there are only four of them - public, thumb, temp, deleted), and
every single one of them will be needed in the future.

Also these buckets can have manually-defined properties or permissions
(e.g. administrator could set those in AWS Management Console),
and all those properties will be lost if the bucket is deleted and recreated.

Change-Id: Ibce052f53eb2827933a63ee1e884df6d30d39864