(feature) Add "privateWiki" configuration key
Crude solution to ensure that images from private wikis won't be leaked.

It's not the best approach in terms of performance (images have to be
downloaded via img_auth.php). A faster solution would be to use S3 signed URLs,
but URLs are formed in FileRepo, not in FileBackend.

Since the private wikis are usually small, using img_auth.php is not a problem.