(config) Add $wgAWSRepoHashLevels (default: 0) for easier migration
This solves issue #17.

When storing images locally, MediaWiki names them a/ab/Filename.png
(where a/ab are hash subdirectores), because local filesystems are slow
when dealing with super-huge directories.

Amazon S3 doesn't have this limitation, so default behavior is to
not add any filename prefix (the objects are named "Filename.png").

However, when the wiki migrates from local storage to Amazon S3
(or vise versa), it's convenient to preserve the directory structure,
which can now be done by setting $wgAWSRepoHashLevels.

$wgAWSRepoHashLevels=2 - use naming [a/b/Filename.png]
$wgAWSRepoHashLevels=0 (default) - use naming [Filename.png]