(one-bucket) Support container paths like "S3BucketName/ExtraPrefix"
This is the first step of replacing four S3 buckets (public, thumb,
deleted and temp) with only one S3 bucket. See issue #18 for details.

Here we allow $wgFileBackends['s3']['containerPaths'] to contain
not just the S3 bucket name, but also an additional (optional) prefix
which is prepended to S3 object names if they belong to this container.

For example, container path "mywiki-img" will mean that "Image.png"
will be uploaded into s3://mywiki-img/Image.png,
however if container path is "mywiki-img/cats/are/funny",
then it will be s3://mywiki-img/cats/are/funny/Image.png.

The purpose is to have the following containerPaths configuration:
[
	'public' => "<BUCKET>",
	'thumb' => "<BUCKET>/thumb",
	'deleted' => "<BUCKET>/deleted",
	'temp' => "<BUCKET>/temp"
]
... where <BUCKET> is the name of only one S3 bucket.
This is similar to how images are stored locally by FSFileBackend.

Implementation details:
1) the code of AmazonS3FileBackend now differentiates between $bucket
(which is the name of S3 bucket) and $container (which is the internal
name of container, as used by MediaWiki).
2) method findContainer() determines $bucket/$prefix from $container.
3) method getBucketAndObject() determines $bucket and S3 object name
from the mwstore:// pseudo-URL (so-called "storage path"),
Methods like doCopyInternal() receive these URLs as parameters.
4) effect of isSecure() is now container-based, not bucket-based.
Presence of .htsecure file is checked at the top of the container,
e.g. in "mywiki/.htsecure" for container path "mywiki"
and in "mywiki/temp/.htsecure" for container path "mywiki/temp".
Convenience method getRestrictFilePath($container) is provided
to calculate the S3 bucket and S3 object name of .htsecure file.

This change keeps backward compatibility.
(wikis that are currently using four S3 buckets will be unaffected)