(tests) Provide testConfigIsLoaded() to decouple the testsuite
Currently the Travis testsuite runs all PHPUnit tests with 2 setups:
old-style configuration and $wgAWSBucketPrefix.
Due to #18 ("one bucket setup" feature), there are now 4 setups.

The testsuite will be decoupled as follows:
1) testConfigIsLoaded() ensures that installBackend() is called at all,
2) testConfig() with configDataProvider() tries all known configuration
variables. It calls installBackend() and then asserts the consequences
(that $wgLocalFileRepo/$wgFileBackends are set to expected values).

After that,
3) AmazonS3FileBackendTest can simply set $wgFileBackends directly
and not worry about how $wgFileBackends gets defined in production,
because this part has already been fully tested by (1) and (2).