(iterator) Add trailing slash to Prefix in ListObjects call from getFileListInternal()
If ListObjects API is called with Prefix without a trailing slash,
CommonPrefixes are returned instead of the list of objects.

For example, when there are objects a/b/1 and a/b/2,
then "ListObjects a/b" will return "CommonPrefixes: a/b/",
while "ListObjects a/b/" will return a/b/1 and a/b/2 (what we need).

Note: we don't append this slash in getS3ListPaginator(),
because getDirectoryListInternal() wants CommonPrefixes.