(tests) Travis: always export BUCKET environment variable
When PHPUnit is called with "--group TestsWithNoNeedForAwsCredentials",
it actually creates a test class for tests that won't be run
(in this case - AmazonS3FileBackendTest).

However, AmazonS3FileBackendTest creates a backend in its constructor,
and than would fail with empty $wgAWSBucketName.

Note: technically, creation of backend should be in setUp().
However, container names are dependent on $wgDBPrefix, which is changed
during the tests (see MediaWikiTestCase::setupDatabaseWithTestPrefix),
leading to "containerPaths" becoming incorrect.
Constructor is called before $wgDBPrefix is modified.