(performance) Don't load entire files into memory in doGetLocalCopyMulti()
This reverts the change of pull request #28.
It was discovered that some wikis store extremely large files that are
still thumbnailable (e.g. PDF files), and this Http::get() could result
in 500Mb. of said files being loaded into the memory.

Instead we return the use of copy() and replace obsolete calls to
wfSuppressWarnings()/wfRestoreWarnings() with local implementation,
just like FSFileBackend does it.