(debug) Log the time spent on actually downloading/uploading S3 objects
This adds a performance log to FileOperations log channel.
Time consumed by the following operations is recorded:

1) downloading S3 object to webserver in doGetLocalCopyMulti(),
2) uploading S3 object from webserver in doCreateInternal(),
3) copying S3 object in doCopyInternal().

This log can be used to troubleshoot performance of Extension:AWS.
It will immediately show if the download/upload is slow because of
bad connectivity between on-premises webserver (not in AWS) and S3.