composer.json: fix version conflict of guzzlehttp/guzzle package
Starting with MediaWiki 1.33, there is a version conflict:
where MediaWiki core installs a fixed version of Guzzle (6.3.3),
but aws/aws-sdk-php also installs Guzzle as dependency, and allows
latter versions of Guzzle to be installed.

In the latest version of Guzzle (6.5.1) this causes issue #29.
("undefined function GuzzleHttp\_idn_uri_convert" error)

Solution is to enforce installation of exactly Guzzle 6.3.3
in our own composer.json.