(bug) Fix error caused by FSFile being returned by doGetLocalCopyMulti
Due to "avoid unnecessary copying" optimization, doGetLocalCopyMulti()
returns FSFile (non-autodeletable file) instead of TempFSFile (which is
normally expected from doGetLocalCopyMulti()).

This causes an issue in UploadBase::getRealPath(), which assumes the
returned value to be TempFSFile, and calls TempFSFile::bind(), a method
that doesn't exist in FSFile.

Solution is to create a fake subclass of FSFile, which pretends to have
the same methods as TempFSFile, but with "do nothing" implementation
(all these methods are related to autodeletion of the file,
but FSFile we are returning shouldn't be deleted).

This is a low-impact bug: getRealPath() was only called when uploading
from stash (e.g. when using "Cancel and return to upload form" button).