(performance) Delay CreateBucket until the error NoSuchBucket is seen
Checking doesBucketExist() on every doPrepareInternal() is unneeded,
because:
1) situation "bucket doesn't exist" happens extremely rarely: either
1a) when Extension:AWS is first installed, or
1b) when bucket was changed in LocalSettings.php, or
1c) bucket was incorrect to begin with (e.g. owned by another AWS user)

2) doPrepareInternal() is called quite often (on every MediaWiki run
that needs to access AmazonS3FileBackend), so it should be optimized
by removing this "almost never useful" API call.

3) Knowing whether the bucket exists is unnecessary, because we receive
NoSuchBucket error from consecutive operations such as PutObject,
and the exception handler of S3Exception can detect this and create
the necessary bucket on the fly. (and then repeat PutObject)

4) {Put,Delete}Object are currently used by do{Publish,Secure}Internal,
so the creation of the bucket would be triggered from doPrepareInternal
even if doPrepareInternal() wouldn't do it itself.