(performance) Avoid many unneeded API calls to {Put,Delete,Head}Object
This is a two-part performance improvement:
1) results of isSecure() are now cached (via Memcached if available).
Most calls to isSecure() won't require a call to S3 API.

2) doPublishInternal() and doSecureInternal() would no longer do S3 API
calls to PutObject and DeleteObject if the bucket is:
- already public (for doPublishInternal)
- already secure (for doSecureInternal)

Main basis for these improvements is that .htsecure file has no effect
on Amazon S3 itself (it's only for internal logic of Extension:AWS),
and serves for persistently remembering "was this container secured?".

Therefore the results of isSecure() check don't need to always check
the presence of .htsecure (nothing except Extension:AWS would be adding
or delething this file) - they can ask a cache instead.

And when Extension:AWS creates/deleted this file (if container security
has changed for whatever reason), we can invalidate the cache in
doPublishInternal() and doSecureInternal(). The result is never stale.

As for not doing PutObject() and DeleteObject() when not necessary,
this is actually a very big optimization, because in 99,9% of wikis
the security of containers never changes (e.g. "deleted" container
will always be secure), meaning that we'll skip many DeleteObject calls
for nonexistent .htsecure in doPublishInternal()
and, similarly, many PutObject() calls for already existing .htsecure.

Note: cache uses "public-read" and "private" strings instead of bools,
because it's not possible to store "false" in the cache.
(false from $cache->get() means "value not found in the cache")